backend/requirements.txt
numpy
pandas
docling
groq
python-multipart
fastapi
uvicorn
pydantic
typing
langchain
langchain-community
python-dotenv